Fix : Separated DTOs into request and response
