+++ src/main/java/com/example/creditmarket/controller/CartController.java
@@ -1,8 +1,8 @@
 package com.example.creditmarket.controller;
 
-import com.example.creditmarket.dto.CartDeleteRequestDTO;
-import com.example.creditmarket.dto.CartResponseDTO;
-import com.example.creditmarket.dto.CartSaveRequestDTO;
+import com.example.creditmarket.dto.request.CartDeleteRequestDTO;
+import com.example.creditmarket.dto.response.CartResponseDTO;
+import com.example.creditmarket.dto.request.CartSaveRequestDTO;
 import com.example.creditmarket.service.cart.CartService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
