+++ src/main/java/com/example/creditmarket/controller/MyPageController.java
@@ -1,7 +1,7 @@
 package com.example.creditmarket.controller;
 
-import com.example.creditmarket.dto.FavoriteResponseDTO;
-import com.example.creditmarket.dto.OrderResponseDTO;
+import com.example.creditmarket.dto.response.FavoriteResponseDTO;
+import com.example.creditmarket.dto.response.OrderResponseDTO;
 import com.example.creditmarket.service.mypage.MyPageService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
