+++ src/main/java/com/example/creditmarket/controller/SearchController.java
@@ -1,6 +1,6 @@
 package com.example.creditmarket.controller;
 
-import com.example.creditmarket.dto.MainListResponseDto;
+import com.example.creditmarket.dto.response.MainListResponseDTO;
 import com.example.creditmarket.service.SearchService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.lang.Nullable;
@@ -19,7 +19,7 @@ public class SearchController {
     private final SearchService searchService;
 
     @GetMapping("/search/results")
-    public List<MainListResponseDto> searchKeyword(@RequestParam(required = false, defaultValue = "") @Nullable String keyword,
+    public List<MainListResponseDTO> searchKeyword(@RequestParam(required = false, defaultValue = "") @Nullable String keyword,
                                                    @RequestParam(required = false, defaultValue = "") @Nullable String loan,
                                                    @RequestParam(required = false, defaultValue = "") @Nullable String age,
                                                    @RequestParam(required = false, defaultValue = "") @Nullable String gender,
