+++ src/main/java/com/example/creditmarket/dto/request/UserLoginRequestDTO.java
@@ -1,4 +1,4 @@
-package com.example.creditmarket.dto;
+package com.example.creditmarket.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
