+++ src/main/java/com/example/creditmarket/service/product/ProductService.java
@@ -1,17 +1,17 @@
 package com.example.creditmarket.service.product;
 
-import com.example.creditmarket.dto.ProductDetailDto;
-import com.example.creditmarket.dto.RecommendResponseDto;
+import com.example.creditmarket.dto.response.ProductDetailResponseDTO;
+import com.example.creditmarket.dto.response.RecommendResponseDTO;
 
 import java.util.List;
 
 public interface ProductService {
 
-    ProductDetailDto getProductDetail(String productId);
+    ProductDetailResponseDTO getProductDetail(String productId);
 
     String buyProduct(String productId,String userEmail);
 
-    List<RecommendResponseDto> recommendList(String userEmail);
+    List<RecommendResponseDTO> recommendList(String userEmail);
 
     String favoriteService(String productId, String userEmail);
 }
