+++ src/main/java/com/example/creditmarket/service/product/impl/ProductServiceImpl.java
@@ -1,6 +1,9 @@
 package com.example.creditmarket.service.product.impl;
 
-import com.example.creditmarket.dto.*;
+import com.example.creditmarket.dto.request.FavoriteRequestDto;
+import com.example.creditmarket.dto.request.OrderRequestDTO;
+import com.example.creditmarket.dto.response.ProductDetailResponseDTO;
+import com.example.creditmarket.dto.response.RecommendResponseDTO;
 import com.example.creditmarket.entity.EntityFProduct;
 import com.example.creditmarket.entity.EntityFavorite;
 import com.example.creditmarket.entity.EntityOption;
@@ -26,10 +29,10 @@ public class ProductServiceImpl implements ProductService {
     /**
      * 상품 상세정보 출력(상품명, 개요, 대상, 한도, 금리 등의 상세정보 출력)
      */
-    public ProductDetailDto getProductDetail(String id) {
+    public ProductDetailResponseDTO getProductDetail(String id) {
         EntityFProduct product = productRepository.findById(id).orElseThrow(()-> new IllegalArgumentException("해당 아이디를 찾을수 없습니다"));
         EntityOption option = optionRepository.findByProductId(id);
-        return new ProductDetailDto(product, option);
+        return new ProductDetailResponseDTO(product, option);
     }
 
     /**
@@ -40,7 +43,7 @@ public class ProductServiceImpl implements ProductService {
                 ()-> new IllegalArgumentException("해당 아이디를 찾을수 없습니다."));
         EntityFProduct product = productRepository.findById(productId).orElseThrow(() ->
                 new IllegalArgumentException("해당 상품을 찾을수 없습니다"));
-        OrderRequestDto orderRequestDto = new OrderRequestDto();
+        OrderRequestDTO orderRequestDto = new OrderRequestDTO();
         orderRequestDto.setUser(user);
         orderRequestDto.setFproduct(product);
         if (orderRepository.save(orderRequestDto.toEntity()) != null) {
@@ -53,14 +56,14 @@ public class ProductServiceImpl implements ProductService {
     /**
      * 추천 게시글
      */
-    public List<RecommendResponseDto> recommendList(String userEmail) {
-        List<RecommendResponseDto> list = new ArrayList<>();
+    public List<RecommendResponseDTO> recommendList(String userEmail) {
+        List<RecommendResponseDTO> list = new ArrayList<>();
         EntityUser user = userRepository.findById(userEmail).orElseThrow();
         List<EntityFProduct> products = productRepository.findProductsByUserPref(user.getUserPrefCreditProductTypeName());
         for (EntityFProduct pr : products) {
             List<EntityOption> options = optionRepository.findOptionByProductIdAndType(pr.getFproduct_id(), user.getUserPrefInterestType());
             for (EntityOption op : options) {
-                list.add(new RecommendResponseDto(pr, op));}}
+                list.add(new RecommendResponseDTO(pr, op));}}
         return list;
     }
